+++ ContentPatcher/Framework/Tokens/ValueProviders/ModConvention/ConventionValueProvider.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Text;
 using Pathoschild.Stardew.Common.Utilities;
@@ -34,7 +33,8 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.ModConvention
         public bool IsMutable => this.Provider.IsMutable();
 
         /// <inheritdoc />
-        public bool IsReady => this.Provider.IsReady();
+        /// <remarks>This is cached to ensure it never changes outside a context update (even if the mod token is otherwise incorrectly changing without a context update), since that would cause subtle hard-to-troubleshoot bugs where patches don't update correctly in some cases.</remarks>
+        public bool IsReady { get; private set; }
 
 
         /*********
@@ -47,6 +47,7 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.ModConvention
         {
             this.Name = name;
             this.Provider = provider;
+            this.IsReady = provider.IsReady();
         }
 
         /// <inheritdoc />
@@ -98,7 +99,12 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.ModConvention
         /// <inheritdoc />
         public bool UpdateContext(IContext context)
         {
-            return this.Provider.UpdateContext();
+            bool wasReady = this.IsReady;
+
+            bool changed = this.Provider.UpdateContext();
+            this.IsReady = this.Provider.IsReady();
+
+            return changed || this.IsReady != wasReady;
         }
 
         /// <inheritdoc />