+++ ContentPatcher/release-notes.md
@@ -1,6 +1,9 @@
 [← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Fixed patches not always updated if they depend on mod-provided tokens that incorrectly change outside a context update.
+
 ## 1.17.1
 Released 19 August 2020.